+++ FormKontrollSystem/Shape.cs
@@ -11,6 +11,6 @@ namespace FormKontrollSystem
         public string Name { get; set; }
         public string Color { get; set; }
 
-        public abstract double CalculateArea();
+        public abstract decimal CalculateArea();
     }
 }
